+++ scrapy/pipelines/__init__.py
