+++ scrapy/pipelines/images.py
@@ -17,8 +17,8 @@ from PIL import Image
 from scrapy.utils.misc import md5sum
 from scrapy.http import Request
 from scrapy.exceptions import DropItem
-#TODO: from scrapy.contrib.pipeline.media import MediaPipeline
-from scrapy.contrib.pipeline.files import FileException, FilesPipeline
+#TODO: from scrapy.pipelines.media import MediaPipeline
+from scrapy.pipelines.files import FileException, FilesPipeline
 
 
 class NoimagesDrop(DropItem):
