+++ scrapy/pipelines/media.py
