+++ tests/test_pipeline_media.py
@@ -8,7 +8,7 @@ from twisted.internet.defer import Deferred, inlineCallbacks
 from scrapy.http import Request, Response
 from scrapy.spider import Spider
 from scrapy.utils.request import request_fingerprint
-from scrapy.contrib.pipeline.media import MediaPipeline
+from scrapy.pipelines.media import MediaPipeline
 from scrapy.utils.signal import disconnect_all
 from scrapy import signals
